Added Authors and Category Controller Admin

+++ src/EDB.WebAPI/Controllers/CategoryController.cs
@@ -1,4 +1,5 @@
-﻿using EDB.Domain.Abstract;
+﻿using EDB.BackofficeUI.Areas.Admin.Models;
+using EDB.Domain.Abstract;
 using EDB.Domain.Concrete;
 using EDB.Domain.Data;
 using EDB.Domain.Entities;
@@ -24,18 +25,19 @@ namespace EDB.WebAPI.Controllers
         }
 
         [HttpGet]
-        [Route("list")]
-        public IActionResult CategoryList()
+        [Route("List")]
+        public IActionResult List()
         {
             var category = _categoryService.List();
             return Ok(category);
         }
         [HttpPost]
-        [Route("add")]
+        [Route("Add")]
         public IActionResult Add(CategoryViewModel category)
         {
             Category cat = new Category()
             {
+                Id = Guid.NewGuid(),
                 Name = category.Name,
             };
             _context.Add(cat);
@@ -43,19 +45,33 @@ namespace EDB.WebAPI.Controllers
             return Ok();
         }
         [HttpPost]
-        [Route("delete")]
-        public IActionResult Delete(Guid id)
+        [Route("Delete")]
+        public IActionResult Delete([FromBody] CategoryRequest response)
         {
-            var category = _categoryService.GetById(id);
+            var category = _categoryService.GetById(response.id);
+            if (category == null)
+            {
+                return NotFound();
+            }
+
             _categoryService.Delete(category);
             return Ok();
         }
         [HttpPost]
-        [Route("update")]
-        public IActionResult Update(Category category) 
-        {     
-            _categoryService.Update(category);
-            return Ok();
+        [Route("Update")]
+        public IActionResult Update(Category category,Guid id) 
+        {
+            var Getid = _categoryService.GetById(id);
+            if (Getid == null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                _categoryService.Update(category);
+                return Ok();
+            }
+          
         }
     }
 }
